refactor(init): split retrofit init modules
