+++ .sampo/changesets/737-split-init-modules.md
@@ -1,0 +1,5 @@
+---
+npm/@wp-typia/project-tools: patch
+---
+
+Split retrofit init planning, package mutations, apply orchestration, and helper script templates into focused runtime modules without changing init behavior.
